rename useAlbumById to useAlbumDoc
