+++ src/hooks/albums/useAlbumDoc.ts
@@ -1,7 +1,7 @@
 import { getAlbumDocById } from "@/data/firestore/albums";
 import { skipToken, useQuery } from "@tanstack/react-query";
 
-export default function useAlbumById(albumId: string | undefined) {
+export default function useAlbumDoc(albumId: string | undefined) {
   return useQuery({
     queryKey: ['albums', albumId],
     queryFn: albumId ? (() => getAlbumDocById(albumId)) : skipToken,
